refactor(proxy): split proxy_tunnel + socket_proxy into focused modules
I never liked these two grab-bag module names. Reorganized by concern:

- proxy_tunnel dissolves: connect_via_socks5 moves into socks5 (all SOCKS5
  in one place), connect_via_http_proxy into a new http_connect module.
- socket_proxy (1026 lines, three distinct implementations) becomes an
  egress/ module directory -- udp (per-flow UDP NAT + SOCKS routing),
  icmp (ICMP echo + checksums), and mod (the dispatcher, renamed
  SocketProxy -> HostEgress). Follows the repo CLAUDE.md rule to split a
  multi-implementation file into a module directory.
- Callers updated: splicetcp egress, arcbox-net (re-exports, datapath,
  example, test), arcbox-vmm, and aroxy's socks/http outbound.

No behavior change. arcbox-proxy 25 tests + clippy green; arcbox-net /
splicetcp / arcbox-vmm check clean.
